gcf_fake_api/requirements.txt
functions-framework==3.*
flask
python-dotenv